agents/datascout_adk/requirements.txt
google-cloud-aiplatform
google-cloud-firestore
adk
fastapi
uvicorn
email-validator
adk
requests
feedparser
